+++ NEWS.md
@@ -7,6 +7,8 @@
     
 ## Minor changes
    
+- Now the `ggpar()` function can handle a list of ggplots.
+
 - New argument `show.legend.text` in the `ggscatter()` function. Use show.legend.text = FALSE to hide text in the legend.
 - New arguments `submain, caption, font.main, font.caption` in the `ggpar()` function.