+++ NEWS.md
@@ -8,7 +8,7 @@
 ## Minor changes
    
 - Now the `ggpar()` function can handle a list of ggplots.
-
+- Now the default legend position is `right`.
 - New argument `show.legend.text` in the `ggscatter()` function. Use show.legend.text = FALSE to hide text in the legend.
 - New arguments `title, submain, subtitle, caption, font.submain, font.subtitle, font.caption` in the `ggpar()` function.