+++ NEWS.md
@@ -2,6 +2,10 @@
 # ggpubr 0.1.1.999
    
     
+## New features
+   
+- New function `get_palette()` to generate a palette of k colors from ggsci palettes, RColorbrewer palettes and custom color palettes. Useful to extend RColorBrewer and ggsci to support more colors.
+  
 ## Minor changes
    
 - Now the `ggpar()` function can handle a list of ggplots.